Refactoring http clients, #1039

+++ EmbyStat.Common/Models/Entities/Helpers/AudioStream.cs
@@ -6,7 +6,7 @@ namespace EmbyStat.Common.Models.Entities.Helpers
     {
         [BsonId]
         public string Id { get; set; }
-        public long? BitRate { get; set; }
+        public int? BitRate { get; set; }
         public string ChannelLayout { get; set; }
         public int? Channels { get; set; }
         public string Codec { get; set; }
