+++ EmbyStat.Common/Models/Entities/Helpers/VideoStream.cs
@@ -8,7 +8,7 @@ namespace EmbyStat.Common.Models.Entities.Helpers
         public string Id { get; set; }
         public string AspectRatio { get; set; }
         public float? AverageFrameRate { get; set; }
-        public long? BitRate { get; set; }
+        public int? BitRate { get; set; }
         public int? Channels { get; set; }
         public int? Height { get; set; }
         public string Language { get; set; }
